# public/openapi.yaml
openapi: 3.0.3
info:
  title: ArtaeFlora HTTP API
  version: 1.0.0
  description: |
    HTTP API of artaeflora.com.

    - **Catalog** endpoints are public reads mirroring exactly what the shop
      shows (active products only).
    - **Admin** endpoints require the `af_admin` session cookie — sign in at
      `/admin` in this browser first, then "Try it out" sends the cookie
      automatically.
    - The website's own pages do NOT use this API — they render server-side.
      Website form mutations (inquiries, admin panel forms) are Next.js server
      actions, which are internal wiring, not part of this contract.
servers:
  - url: /
    description: This host (wherever these docs are served from)
  - url: http://localhost:3000
    description: Local development
  - url: https://artaeflora.com
    description: Production

tags:
  - name: catalog
    description: Public product/category reads (active products only)
  - name: checkout
    description: Cart to Stripe Checkout session
  - name: webhooks
    description: Inbound calls from Stripe
  - name: admin
    description: Requires the admin session cookie

paths:
  /api/products:
    get:
      tags: [catalog]
      summary: List active products
      operationId: listProducts
      security: []
      parameters:
        - name: category
          in: query
          schema: { type: string }
          description: Category slug (parent categories include their subcategories).
          example: candles
        - name: occasion
          in: query
          schema: { type: string }
          example: Diwali
        - name: q
          in: query
          schema: { type: string }
          description: Free-text search over name, description, and occasions.
      responses:
        "200":
          description: Matching products.
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  products:
                    type: array
                    items: { $ref: "#/components/schemas/Product" }
        "404":
          description: Unknown category slug.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /api/products/{slug}:
    get:
      tags: [catalog]
      summary: Get one product by slug
      operationId: getProduct
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
          example: lavender-soy-candle
      responses:
        "200":
          description: The product.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "404":
          description: Not found or not active.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /api/categories:
    get:
      tags: [catalog]
      summary: Category tree with product counts
      operationId: listCategories
      security: []
      responses:
        "200":
          description: Top-level categories with nested subcategories.
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: array
                    items: { $ref: "#/components/schemas/CategoryNode" }

  /api/admin/products:
    get:
      tags: [admin]
      summary: List ALL products (including hidden)
      operationId: adminListProducts
      security: [{ adminCookie: [] }]
      responses:
        "200":
          description: Every product with admin fields (active flag, ids).
        "401":
          description: Not signed in.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    post:
      tags: [admin]
      summary: Create or update a product
      description: >-
        Include `id` to update an existing product; omit it to create.
        Category can be given as `categoryId` or `categorySlug`.
        On update, only the fields you send change; sending `images`
        replaces the full image list.
      operationId: adminSaveProduct
      security: [{ adminCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminProductInput" }
            example:
              name: Marigold Garland Candle
              description: Hand-poured candle wrapped in a clay marigold garland.
              categorySlug: festive-candles
              priceCents: 2200
              occasions: [Diwali, Weddings]
              featured: false
              active: true
              images: [{ url: /products/candles.png }]
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SaveResult" }
        "200":
          description: Updated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SaveResult" }
        "400":
          description: Validation problem (missing fields, bad price, unknown category).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Not signed in.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Update requested for an id that doesn't exist.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /api/admin/categories:
    get:
      tags: [admin]
      summary: Flat category list with ids
      operationId: adminListCategories
      security: [{ adminCookie: [] }]
      responses:
        "200":
          description: All categories with parent info and product counts.
        "401":
          description: Not signed in.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    post:
      tags: [admin]
      summary: Create or update a category
      description: Include `id` to update; omit to create.
      operationId: adminSaveCategory
      security: [{ adminCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id: { type: string, description: Present = update }
                name: { type: string }
                slug: { type: string }
                parentId: { type: string, nullable: true }
                sortOrder: { type: integer }
            example: { name: Wall Art, sortOrder: 50 }
      responses:
        "201": { description: Created. }
        "200": { description: Updated. }
        "400":
          description: Missing name on create.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Not signed in.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Unknown id on update.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /api/admin/orders:
    get:
      tags: [admin]
      summary: List all orders with items
      operationId: adminListOrders
      security: [{ adminCookie: [] }]
      responses:
        "200":
          description: Orders, newest first, including line items, customer contact, and shipping address.
        "401":
          description: Not signed in.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /api/admin/inquiries:
    get:
      tags: [admin]
      summary: List inquiries
      operationId: adminListInquiries
      security: [{ adminCookie: [] }]
      parameters:
        - name: type
          in: query
          schema: { type: string, enum: [PAINTING, EVENT, CLASS, GENERAL] }
        - name: handled
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: Matching inquiries, newest first.
        "401":
          description: Not signed in.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /api/checkout:
    post:
      tags: [checkout]
      summary: Create a checkout for the current cart
      description: |
        Validates every cart line against the database (product must exist, be
        active, not inquiry-only, and have a price — client-sent prices are
        never trusted) and computes shipping from the store settings.

        Behavior depends on configuration:
        - `STRIPE_SECRET_KEY` set → creates a Stripe Checkout Session and
          returns its hosted-page URL.
        - Key missing in **development** → mock mode: records the order
          immediately and returns a local success URL.
        - Key missing in **production** → 503 (never gives products away).
      operationId: createCheckout
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
            example:
              items:
                - productId: cml8x2f0q0003abcd
                  quantity: 2
              fulfillment: SHIPPING
      responses:
        "200":
          description: Redirect target for the buyer's browser.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "400":
          description: Malformed body or empty cart.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: >-
            One or more items are unknown, inactive, inquiry-only, or
            unpriced — the client should refresh its cart.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Production instance without Stripe configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/webhooks/stripe:
    post:
      tags: [webhooks]
      summary: Stripe event receiver
      description: |
        Called by Stripe, never by browsers. The raw body is verified against
        `Stripe-Signature` using the webhook signing secret. On
        `checkout.session.completed` an Order (+ items, customer contact,
        shipping address, shipping cost) is recorded. Idempotent: replayed
        events for an already-recorded session are acknowledged and ignored.
      operationId: stripeWebhook
      security: []
      parameters:
        - name: Stripe-Signature
          in: header
          required: true
          schema:
            type: string
          description: HMAC signature set by Stripe.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Raw Stripe event payload (verified before parsing).
      responses:
        "200":
          description: Event acknowledged.
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
        "400":
          description: Missing or invalid signature.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Stripe env vars not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/upload:
    post:
      tags: [admin]
      summary: Upload a product/gallery/hero media file
      description: |
        Admin-only (session cookie). Stores to Cloudinary when
        `CLOUDINARY_URL` is set, otherwise to `public/uploads/<folder>/` with a
        collision-proof generated filename. Video is accepted only for the
        `hero` folder. Limits: images 8 MB, video 60 MB.
      operationId: uploadMedia
      security:
        - adminCookie: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, folder]
              properties:
                file:
                  type: string
                  format: binary
                  description: JPEG, PNG, WebP, GIF, MP4, or WebM.
                folder:
                  type: string
                  enum: [products, gallery, hero]
      responses:
        "200":
          description: Stored; use the returned URL in the entity being saved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
        "400":
          description: No file, invalid folder, or video outside the hero folder.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid admin session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: File exceeds the size limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "415":
          description: Unsupported media type.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  securitySchemes:
    adminCookie:
      type: apiKey
      in: cookie
      name: af_admin
      description: >-
        HMAC-SHA256-signed session token set by the admin login form
        (7-day expiry). Obtainable only via /admin/login.
  schemas:
    CheckoutRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [productId, quantity]
            properties:
              productId:
                type: string
                description: Product id (cuid) from the catalog.
              quantity:
                type: integer
                minimum: 1
                maximum: 99
        fulfillment:
          type: string
          enum: [PICKUP, SHIPPING]
          default: PICKUP
    CheckoutResponse:
      type: object
      required: [url]
      properties:
        url:
          type: string
          description: >-
            Absolute Stripe-hosted payment page URL, or a site-relative
            success path in development mock mode.
    UploadResponse:
      type: object
      required: [url, mediaType]
      properties:
        url:
          type: string
          example: /uploads/products/6f0d0f6e-....jpg
        mediaType:
          type: string
          enum: [IMAGE, VIDEO]
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Product:
      type: object
      properties:
        id: { type: string }
        slug: { type: string }
        name: { type: string }
        description: { type: string }
        priceCents:
          type: integer
          nullable: true
          description: Null for inquiry-only products.
        inquiryOnly: { type: boolean }
        occasions:
          type: array
          items: { type: string }
        featured: { type: boolean }
        category: { $ref: "#/components/schemas/CategoryRef" }
        images:
          type: array
          items:
            type: object
            properties:
              url: { type: string }
              alt: { type: string, nullable: true }
    CategoryRef:
      type: object
      properties:
        id: { type: string }
        slug: { type: string }
        name: { type: string }
        parent:
          type: object
          nullable: true
          properties:
            id: { type: string }
            slug: { type: string }
            name: { type: string }
    CategoryNode:
      type: object
      properties:
        id: { type: string }
        slug: { type: string }
        name: { type: string }
        productCount: { type: integer }
        subcategories:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              slug: { type: string }
              name: { type: string }
              productCount: { type: integer }
    AdminProductInput:
      type: object
      properties:
        id: { type: string, description: Present = update that product }
        name: { type: string }
        slug: { type: string, description: Optional — generated from name }
        description: { type: string }
        priceCents:
          type: integer
          nullable: true
          description: Cents (1800 = $18.00). Ignored when inquiryOnly is true.
        inquiryOnly: { type: boolean, default: false }
        occasions:
          type: array
          items: { type: string }
        featured: { type: boolean }
        active: { type: boolean }
        categoryId: { type: string }
        categorySlug: { type: string, description: Alternative to categoryId }
        images:
          type: array
          description: Full replacement list; first image is the main one.
          items:
            type: object
            required: [url]
            properties:
              url: { type: string }
              alt: { type: string }
    SaveResult:
      type: object
      properties:
        id: { type: string }
        slug: { type: string }
        created: { type: boolean }
        updated: { type: boolean }